fix(ks): add required cookie groups to all cookie settings pages
Refs: YJDH-977 (Add required cookie groups to cookie settings pages)
Refs: YJDH-969 (PR City-of-Helsinki/yjdh#4242 added required cookies)
