+++ frontend/kesaseteli/employer/src/pages/_app.tsx
@@ -4,12 +4,12 @@ import 'hds-design-tokens';
 import AuthProvider from 'kesaseteli/employer/auth/AuthProvider';
 import Footer from 'kesaseteli/employer/components/footer/Footer';
 import Header from 'kesaseteli/employer/components/header/Header';
+import getRequiredEmployerCookieGroups from 'kesaseteli/employer/utils/get-required-employer-cookie-groups';
 import { getBackendDomain } from 'kesaseteli-shared/backend-api/backend-api';
 import { COOKIE_CONSENT_SITE_NAME } from 'kesaseteli-shared/constants/cookie-consent';
 import { ROUTES } from 'kesaseteli-shared/constants/routes';
 import useMatomo from 'kesaseteli-shared/hooks/useMatomo';
 import createQueryClient from 'kesaseteli-shared/query-client/create-query-client';
-import getRequiredCookieGroups from 'kesaseteli-shared/utils/getRequiredCookieGroups';
 import { AppProps } from 'next/app';
 import dynamic from 'next/dynamic';
 import { useRouter } from 'next/router';
@@ -21,19 +21,11 @@ import BaseApp from 'shared/components/app/BaseApp';
 import ConfirmDialog from 'shared/components/confirm-dialog/ConfirmDialog';
 import Portal from 'shared/components/confirm-dialog/Portal';
 import { DialogContextProvider } from 'shared/contexts/DialogContext';
-import { type RequiredGroups } from 'shared/utils/cookieConsentSettings';
 
 const CookieConsent = dynamic(
   () => import('kesaseteli-shared/components/cookieConsent/CookieConsent'),
   { ssr: false }
 );
-
-const getRequiredEmployerCookieGroups = (): RequiredGroups =>
-  getRequiredCookieGroups({
-    includeSessionIdCookie: true, // Needed because of Django's session handling
-    includeSamlSessionCookie: true, // Needed because of Suomi.fi login
-    includeCsrfTokenCookie: true, // Needed because of POSTing data to backend
-  });
 
 const App: React.FC<AppProps> = (appProps) => {
   const isMatomoConfigured = useMatomo();
