+++ frontend/kesaseteli/employer/src/pages/cookie-settings.tsx
@@ -1,3 +1,4 @@
+import getRequiredEmployerCookieGroups from 'kesaseteli/employer/utils/get-required-employer-cookie-groups';
 import CookieSettings from 'kesaseteli-shared/components/cookieSettings/CookieSettings';
 import { GetStaticProps, NextPage } from 'next';
 import { useTranslation } from 'next-i18next';
@@ -11,6 +12,7 @@ const CookieSettingsPage: NextPage = () => {
     <CookieSettings
       title={`${t('common:appName')} - ${t('common:cookieSettings')}`}
       siteName={t('common:appName')}
+      requiredGroups={getRequiredEmployerCookieGroups()}
     />
   );
 };
