+++ frontend/kesaseteli/youth/src/pages/_app.tsx
@@ -3,12 +3,12 @@ import 'hds-design-tokens';
 
 import Footer from 'kesaseteli/youth/components/footer/Footer';
 import Header from 'kesaseteli/youth/components/header/Header';
+import getRequiredYouthCookieGroups from 'kesaseteli/youth/utils/get-required-youth-cookie-groups';
 import { getBackendDomain } from 'kesaseteli-shared/backend-api/backend-api';
 import { COOKIE_CONSENT_SITE_NAME } from 'kesaseteli-shared/constants/cookie-consent';
 import { ROUTES } from 'kesaseteli-shared/constants/routes';
 import useMatomo from 'kesaseteli-shared/hooks/useMatomo';
 import createQueryClient from 'kesaseteli-shared/query-client/create-query-client';
-import getRequiredCookieGroups from 'kesaseteli-shared/utils/getRequiredCookieGroups';
 import { AppProps } from 'next/app';
 import dynamic from 'next/dynamic';
 import { useRouter } from 'next/router';
@@ -17,7 +17,6 @@ import React from 'react';
 import { QueryClientProvider } from 'react-query';
 import BackendAPIProvider from 'shared/backend-api/BackendAPIProvider';
 import BaseApp from 'shared/components/app/BaseApp';
-import { type RequiredGroups } from 'shared/utils/cookieConsentSettings';
 
 const CookieConsent = dynamic(
   () => import('kesaseteli-shared/components/cookieConsent/CookieConsent'),
@@ -25,9 +24,6 @@ const CookieConsent = dynamic(
 );
 
 const queryClient = createQueryClient();
-
-const getRequiredYouthCookieGroups = (): RequiredGroups =>
-  getRequiredCookieGroups();
 
 const App: React.FC<AppProps> = (appProps: AppProps) => {
   const isMatomoConfigured = useMatomo();
