+++ frontend/kesaseteli/youth/src/utils/get-required-youth-cookie-groups.ts
@@ -1,0 +1,7 @@
+import getRequiredCookieGroups from 'kesaseteli-shared/utils/getRequiredCookieGroups';
+import type { RequiredGroups } from 'shared/utils/cookieConsentSettings';
+
+const getRequiredYouthCookieGroups = (): RequiredGroups =>
+  getRequiredCookieGroups();
+
+export default getRequiredYouthCookieGroups;
